requirements.txt
numpy
scipy
tensorflow-gpu==1.14
sklearn
Pillow
matplotlib
jupyter
tqdm